+++ src/main/java/org/replikativ/kabel/MessageHandlerBinary.java
@@ -1,4 +1,4 @@
-package io.replikativ.kabel;
+package org.replikativ.kabel;
 
 import javax.websocket.MessageHandler;
 import java.nio.ByteBuffer;